Remove FetchResources for the DistIndex source
Instead replace it with a comment that the `aws` cli utility should be used to obtain the input data.

The next commit will contain a new Source which re-adds the FetchResources implementation. It has moved because the input data we obtain is different from the input data obtained from the aws utility, and requires a different way to parse it. By adding a separate source we keep both options open.