Split backends to separate crates
Aims reduce the required dependencies, when not all backends are necessary

The library is split in seven parts.

First a core crate which contains everything necessary
to create an index of releases.
Secondly, an i/o crate which optionally includes an HTTP client; and is used as a helper crate for some of the source implementations.
Then, the four source implementations have each their own crate.
And lastly, the original `rust-releases` source implementation which re-exports these crates, and provides Cargo features (conditional compilation), through which source implementations may be enabled and disabled at will.