Replace rusoto with aws-sdk
Rusoto is unmaintained, and relies on outdated dependencies which RustSec reports as insecure.

The Rust AWS sdk is still in preview, but weighted against the above, already a better choice.

Unfortunately, no anonymous credential provider is available (like Rusoto or the Java AWS SDK).
We have to use a workaround instead, where we manually create the S3 operation using a lower level API, make the operation unsigned, and execute the operation using the lower level Smithy client.