Add new base crates for complete rework
Added rust-releases-core2 which will eventually replace rust-releases-core.
This library contains the primary "Release" structs while everything "toolchain"
related will be moved to the rust-toolchain crate.

The second added crate is rust-releases-compact-index; for this crate the intention
is to create a compact version from AWS index which is too large to be pulled from
quickly.